move bp_sufficiency to own file
